# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Football Match Interaction Platform API
  version: 1.0.0
servers:
  - url: http://localhost:7001
...
paths:
...
  /api/competitions:
    get:
      tags: [Competitions]
      operationId: listCompetitions
      summary: 获取赛事列表
      security: []
      x-spec:
        file: specs/002-competition-stage-match.md
        acceptanceCriteria: [AC-01, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "200":
          description: 赛事列表，按创建顺序升序返回；无数据时 `data` 为空数组。
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/competitions/{competitionId}:
    get:
      tags: [Competitions]
      operationId: getCompetition
      summary: 获取赛事详情
      security: []
      x-spec:
        file: specs/002-competition-stage-match.md
        acceptanceCriteria: [AC-02, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/CompetitionId"
      responses:
        "200":
          description: 赛事详情
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/stages/{stageId}:
    get:
      tags: [Stages]
      operationId: getStage
      summary: 获取阶段详情
      security: []
      x-spec:
        file: specs/002-competition-stage-match.md
        acceptanceCriteria: [AC-02, AC-05]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/StageId"
      responses:
        "200":
          description: 阶段详情
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StageResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/teams:
    get:
      tags: [Teams]
      operationId: listTeams
      summary: 获取球队列表
      security: []
      x-spec:
        file: specs/003-teams.md
        acceptanceCriteria: [AC-01, AC-05]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "200":
          description: 球队列表，按名称升序和 ID 升序返回；无数据时 `data` 为空数组。
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/teams/{teamId}:
    get:
      tags: [Teams]
      operationId: getTeam
      summary: 获取球队详情
      security: []
      x-spec:
        file: specs/003-teams.md
        acceptanceCriteria: [AC-02, AC-04]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: 球队详情，只包含本地权威资料。
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/matches:
    get:
      tags: [Matches]
      operationId: listMatches
      summary: 获取比赛列表
      security: []
      x-spec:
        file: specs/002-competition-stage-match.md
        acceptanceCriteria: [AC-03, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - name: competitionId
          in: query
          required: false
          description: 按赛事筛选比赛。
          schema:
            type: integer
            minimum: 1
        - name: stageId
          in: query
          required: false
          description: 按阶段筛选比赛。
          schema:
            type: integer
            minimum: 1
        - name: status
          in: query
          required: false
          description: 按比赛状态筛选。
          schema:
            $ref: "#/components/schemas/MatchStatus"
      responses:
        "200":
          description: 比赛列表，默认按开赛时间升序和 ID 升序返回；无匹配时 `data` 为空数组。
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/matches/{matchId}:
    get:
      tags: [Matches]
      operationId: getMatch
      summary: 获取比赛详情
      security: []
      x-spec:
        file: specs/002-competition-stage-match.md
        acceptanceCriteria: [AC-04, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/MatchId"
      responses:
        "200":
          description: 比赛详情，包含赛事、阶段、球队、状态和最新比赛结果。
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/admin/competitions:
    post:
      tags: [Admin]
      operationId: createCompetition
      summary: 创建赛事
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-01, AC-04, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        $ref: "#/components/requestBodies/CompetitionWriteRequest"
      responses:
        "201":
          description: 创建成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/competitions/{competitionId}:
    put:
      tags: [Admin]
      operationId: updateCompetition
      summary: 编辑赛事
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-02, AC-04, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/CompetitionId"
      requestBody:
        $ref: "#/components/requestBodies/CompetitionWriteRequest"
      responses:
        "200":
          description: 编辑成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Admin]
      operationId: deleteCompetition
      summary: 删除赛事
      description: 删除 Competition 时级联删除其 Stage 和 Match；用户侧查询不再返回这些基础数据。
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-03, AC-04, AC-05]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/CompetitionId"
      responses:
        "204":
          description: 删除成功，无响应体。
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/stages:
    post:
      tags: [Admin]
      operationId: createStage
      summary: 创建赛事阶段
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-01, AC-04, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        $ref: "#/components/requestBodies/StageWriteRequest"
      responses:
        "201":
          description: 创建成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StageResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/stages/{stageId}:
    put:
      tags: [Admin]
      operationId: updateStage
      summary: 编辑赛事阶段
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-02, AC-04, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/StageId"
      requestBody:
        $ref: "#/components/requestBodies/StageWriteRequest"
      responses:
        "200":
          description: 编辑成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StageResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Admin]
      operationId: deleteStage
      summary: 删除赛事阶段
      description: 删除 Stage 时级联删除其 Match；用户侧查询不再返回这些基础数据。
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-03, AC-04, AC-05]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/StageId"
      responses:
        "204":
          description: 删除成功，无响应体。
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/teams:
    post:
      tags: [Admin]
      operationId: createTeam
      summary: 创建球队
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-01, AC-04, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        $ref: "#/components/requestBodies/TeamWriteRequest"
      responses:
        "201":
          description: 创建成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/teams/{teamId}:
    put:
      tags: [Admin]
      operationId: updateTeam
      summary: 编辑球队
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-02, AC-04, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        $ref: "#/components/requestBodies/TeamWriteRequest"
      responses:
        "200":
          description: 编辑成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Admin]
      operationId: deleteTeam
      summary: 删除球队
      description: 删除 Team 时级联删除该队参与的 Match；用户侧查询不再返回这些基础数据。
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-03, AC-04, AC-05]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/TeamId"
      responses:
        "204":
          description: 删除成功，无响应体。
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/matches:
    post:
      tags: [Admin]
      operationId: createMatch
      summary: 创建比赛
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-01, AC-04, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        $ref: "#/components/requestBodies/MatchWriteRequest"
      responses:
        "201":
          description: 创建成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/matches/{matchId}:
    put:
      tags: [Admin]
      operationId: updateMatch
      summary: 编辑比赛
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-02, AC-04, AC-05, AC-06]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/MatchId"
      requestBody:
        $ref: "#/components/requestBodies/MatchWriteRequest"
      responses:
        "200":
          description: 编辑成功
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Admin]
      operationId: deleteMatch
      summary: 删除比赛
      description: 删除 Match 后，用户侧比赛查询不再返回该比赛。
      security:
        - cookieAuth: []
      x-spec:
        file: specs/010-admin-data-management.md
        acceptanceCriteria: [AC-03, AC-04, AC-05]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/MatchId"
      responses:
        "204":
          description: 删除成功，无响应体。
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: 服务端会话 Cookie。凭据不得放入 Query、响应体、错误详情或日志。
  parameters:
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      description: 可选请求追踪 ID；缺失或格式不合法时由服务端生成。
      schema:
        type: string
        maxLength: 100
        pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,99}$"
    CompetitionId:
      name: competitionId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    StageId:
      name: stageId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    MatchId:
      name: matchId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  headers:
    RequestId:
      description: 本次请求的追踪 ID。
      schema:
        type: string
        minLength: 1
  requestBodies:
    CompetitionWriteRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CompetitionWriteRequest"
    StageWriteRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StageWriteRequest"
    TeamWriteRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TeamWriteRequest"
    MatchWriteRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MatchWriteRequest"
...
  responses:
    ValidationFailed:
      description: 请求输入不符合契约或业务校验规则。
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未认证或会话无效。
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 已认证但没有权限执行该操作。
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 请求的资源不存在。
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: 服务端发生未预期错误；响应不得包含 SQL、堆栈、密钥或内部路径。
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    StageType:
      type: string
      enum: [GROUP, LEAGUE, KNOCKOUT]
    MatchStatus:
      type: string
      enum: [SCHEDULED, LIVE, FINISHED]
      description: 分别表示未开始、进行中、已结束。
...
    Competition:
      type: object
      additionalProperties: false
      required: [id, name, description, createdAt, updatedAt]
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 2
          maxLength: 80
        description:
          type: string
          minLength: 0
          maxLength: 500
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CompetitionWriteRequest:
      type: object
      additionalProperties: false
      required: [name, description]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
          description: 服务端去除首尾空白后校验。
        description:
          type: string
          minLength: 0
          maxLength: 500
          description: 服务端去除首尾空白后保存。
    CompetitionResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Competition"
    CompetitionListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Competition"
    Stage:
      type: object
      additionalProperties: false
      required:
        [
          id,
          competitionId,
          name,
          type,
          groupName,
          sortOrder,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
          minimum: 1
        competitionId:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 2
          maxLength: 80
        type:
          $ref: "#/components/schemas/StageType"
        groupName:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 40
          description: GROUP 阶段可用于区分小组；LEAGUE 和 KNOCKOUT 通常为 null。
        sortOrder:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StageWriteRequest:
      type: object
      additionalProperties: false
      required: [competitionId, name, type, groupName, sortOrder]
      properties:
        competitionId:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 2
          maxLength: 80
          description: 服务端去除首尾空白后校验。
        type:
          $ref: "#/components/schemas/StageType"
        groupName:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 40
        sortOrder:
          type: integer
          minimum: 0
    StageResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Stage"
    Team:
      type: object
      additionalProperties: false
      required:
        [id, name, shortName, logoUrl, openLigaDbTeamId, createdAt, updatedAt]
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
          minLength: 2
          maxLength: 80
        shortName:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 30
        logoUrl:
          type:
            - string
            - "null"
          format: uri
          maxLength: 500
        openLigaDbTeamId:
          type:
            - integer
            - "null"
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TeamWriteRequest:
      type: object
      additionalProperties: false
      required: [name, shortName, logoUrl, openLigaDbTeamId]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
          description: 服务端去除首尾空白后校验。
        shortName:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 30
        logoUrl:
          type:
            - string
            - "null"
          format: uri
          maxLength: 500
        openLigaDbTeamId:
          type:
            - integer
            - "null"
          minimum: 1
    TeamResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Team"
    TeamListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Team"
...
    MatchResult:
      type: object
      additionalProperties: false
      required: [homeScore, awayScore, updatedAt]
      properties:
        homeScore:
          type: integer
          minimum: 0
        awayScore:
          type: integer
          minimum: 0
        updatedAt:
          type: string
          format: date-time
    Match:
      type: object
      additionalProperties: false
      required:
        [
          id,
          competition,
          stage,
          homeTeam,
          awayTeam,
          startsAt,
          status,
          groupName,
          knockoutRound,
          bracketPosition,
          result,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
          minimum: 1
        competition:
          $ref: "#/components/schemas/Competition"
        stage:
          $ref: "#/components/schemas/Stage"
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        startsAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/MatchStatus"
        groupName:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 40
        knockoutRound:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 40
        bracketPosition:
          type:
            - integer
            - "null"
          minimum: 1
        result:
          anyOf:
            - $ref: "#/components/schemas/MatchResult"
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MatchWriteRequest:
      type: object
      additionalProperties: false
      required:
        [
          stageId,
          homeTeamId,
          awayTeamId,
          startsAt,
          status,
          groupName,
          knockoutRound,
          bracketPosition,
        ]
      properties:
        stageId:
          type: integer
          minimum: 1
        homeTeamId:
          type: integer
          minimum: 1
        awayTeamId:
          type: integer
          minimum: 1
        startsAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/MatchStatus"
        groupName:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 40
        knockoutRound:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 40
        bracketPosition:
          type:
            - integer
            - "null"
          minimum: 1
...
    MatchResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Match"
    MatchListResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Match"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
        requestId:
          type: string
          minLength: 1